requirements.txt
altair==6.0.0
anyio==4.12.1
asn1crypto==1.5.1
attrs==25.4.0
blinker==1.9.0
boto3==1.42.51
boto3-stubs==1.42.51
botocore==1.42.51
botocore-stubs==1.42.41
cachetools==7.0.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
filelock==3.24.3
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mypy-boto3-s3==1.42.37
narwhals==2.17.0
numpy==2.4.2
packaging==26.0
pandas==2.3.3
pillow==12.1.1
platformdirs==4.9.2
pluggy==1.6.0
protobuf==6.33.5
pyarrow==23.0.1
pycparser==3.0
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==25.3.0
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
smmap==5.0.2
snowflake-connector-python==4.3.0
sortedcontainers==2.4.0
streamlit==1.55.0
structlog==25.5.0
tenacity==9.1.4
toml==0.10.2
tomlkit==0.14.0
tornado==6.5.4
types-awscrt==0.31.2
types-s3transfer==0.16.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
